WOR-433 Split app/core/metrics.py into models / store / tags
The 864-LOC metrics.py mixed three independent concerns: Pydantic row
models, the SQLite-backed MetricsStore, and the auto-tag rules. Splitting
along that natural seam lets reporting / analytics callers import the
row models without dragging in the store, and shrinks every file under
the 700-LOC RECOMMEND threshold.

New layout:

  app/core/metrics_models.py    329 LOC  Pydantic row models +
                                          read-side aggregates.
  app/core/metrics_store.py     565 LOC  MetricsStore class, SQL CREATE
                                          + ALTER constants, _row_to_metrics.
  app/core/metrics_tags.py       98 LOC  compute_tags + _anomaly_tags +
                                          _category_tags helpers.
  app/core/metrics.py            42 LOC  Facade — re-exports the public
                                          surface for back-compat.

The facade preserves every name that test fixtures and tooling currently
import from `app.core.metrics`, so no test or call-site change is needed.
Callers can migrate to direct imports incrementally.

New import-linter contract `metrics-models-no-store` enforces the one
boundary that the split exists to create: row models must not depend
on the store, otherwise the back-edge would re-tangle the modules.

All 1601 tests pass. lint-imports: 8 contracts kept, 0 broken.
Closes WOR-433
